refactored due to feedback
