+++ fetcher-web/src/main/kotlin/org/bibletranslationtools/fetcher/impl/repository/RCRepositoryImpl.kt
@@ -3,6 +3,7 @@ package org.bibletranslationtools.fetcher.impl.repository
 import org.bibletranslationtools.fetcher.repository.ResourceContainerRepository
 import org.bibletranslationtools.fetcher.repository.SourceCacheAccessor
 import org.bibletranslationtools.fetcher.repository.StorageAccess
+import org.slf4j.LoggerFactory
 import java.io.File
 
 
@@ -10,6 +11,7 @@ class RCRepositoryImpl(
     private val storageAccess: StorageAccess,
     private val sourceCacheAccessor: SourceCacheAccessor
 ) : ResourceContainerRepository {
+    private val logger = LoggerFactory.getLogger(javaClass)
     private val rcTemplateName = "%s_%s"
 
     override fun getRC(
@@ -29,8 +31,19 @@ class RCRepositoryImpl(
             resourceId
         )?.let { repoUrl ->
             val repoName = "${languageCode}_$resourceId"
+            val reposDirectory = storageAccess.getReposDir()
 
-            sourceCacheAccessor.downloadRepo(repoName, repoUrl)
+            val process = ProcessBuilder()
+                .command("git", "clone", repoUrl, repoName)
+                .directory(reposDirectory)
+                .start()
+
+            val exit = process.waitFor()
+
+            if (exit != 0) {
+                logger.error("An error occurred in cloneRepo with exit code: $exit")
+            }
+
             storageAccess.getRepoFromFileSystem(repoName)
         }
     }
