+++ fetcher-web/src/main/kotlin/org/bibletranslationtools/fetcher/repository/SourceCacheAccessor.kt
@@ -3,5 +3,4 @@ package org.bibletranslationtools.fetcher.repository
 interface SourceCacheAccessor {
     fun update()
     fun getRepoUrl(languageCode: String, resourceId: String): String?
-    fun downloadRepo(repoName: String, repoUrl: String)
 }
